requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.10.0
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
backports.functools-lru-cache==2.0.0
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.0
constantly==23.10.4
cryptography==46.0.3
cssselect2==0.8.0
daphne==4.2.1
dj-database-url==3.0.1
Django==4.2.25
djangorestframework==3.16.1
fonttools==4.60.1
frozenlist==1.8.0
git-filter-repo==2.47.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
indic-transliteration==2.3.75
markdown-it-py==4.0.0
mdurl==0.1.2
msgpack==1.1.2
multidict==6.7.0
packaging==25.0
pillow==12.0.0
propcache==0.4.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyphen==0.17.2
python-dotenv==1.1.1
pytube==15.0.0
redis==7.0.0
regex==2025.10.23
reportlab==4.4.4
requests==2.32.5
rich==14.2.0
roman==5.1
service-identity==24.2.0
setuptools==80.9.0
shellingham==1.5.4
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
twilio==9.8.4
Twisted==25.5.0
txaio==25.9.2
typer==0.20.0
typing_extensions==4.15.0
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.11.0
yarl==1.22.0
yt-dlp==2025.10.18.232824.dev0
zope.interface==8.0.1
zopfli==0.2.3.post1
